Создает маршрут Cities
