+++ src/components/app/app.tsx
@@ -1,11 +1,21 @@
-import MainScreen from '../../pages/main/main';
+import { BrowserRouter, Routes, Route } from 'react-router-dom';
+import Layout from '../../pages/layout/layout';
+import Cities from '../../pages/cities/cities';
 
 type AppProps = {
   numberOffers: number;
 };
 
 function App({ numberOffers }: AppProps): JSX.Element {
-  return <MainScreen numberOffers={numberOffers} />;
+  return (
+    <BrowserRouter>
+      <Routes>
+        <Route path="/" element={<Layout />}>
+          <Route index element={<Cities numberOffers={numberOffers} />} />
+        </Route>
+      </Routes>
+    </BrowserRouter>
+  );
 }
 
 export default App;
